backend/requirements.txt
fastapi==0.100.0
uvicorn==0.24.0
requests==2.31.0
types-requests==2.31.0
python-dotenv==1.0.0
spotipy==2.25.1
openai>=1.0.0
Pillow>=10.0.0






